Add network configuration support for synchronous file uploads
Introduce RemoteNetworkConfigSerializer to centralize shared network
settings (timeouts, proxies, and certificates) and ensure consistent
validation across persistent Remotes and ad-hoc upload downloads.

- Exclude sync-specific fields (policy, rate_limit) from shared config
- Pass validated network kwargs to the downloader in UploadSerializerFieldsMixin
- Reuse common proxy and certificate validation logic

closes: #7201
