+++ CHANGES/7201.feature
@@ -1,0 +1,1 @@
+Add network configuration support for synchronous file uploads
